QuestionsStorage moved to its own package
