+++ src/main/java/ru/hh/resumebuilderbot/message/handler/AnswerMessageHandler.java
@@ -2,12 +2,12 @@ package ru.hh.resumebuilderbot.message.handler;
 
 import ru.hh.resumebuilderbot.Answer;
 import ru.hh.resumebuilderbot.CurrentUserState;
-import ru.hh.resumebuilderbot.QuestionsStorage;
 import ru.hh.resumebuilderbot.question.generator.FixedQuestionGenerator;
 import ru.hh.resumebuilderbot.question.generator.QuestionGeneratorByNumber;
 import ru.hh.resumebuilderbot.question.generator.QuestionGeneratorsQueue;
 import ru.hh.resumebuilderbot.texts.storage.TextId;
 import ru.hh.resumebuilderbot.user.data.storage.UserDataStorage;
+import ru.hh.resumebuilderbot.questions.storage.QuestionsStorage;
 
 public class AnswerMessageHandler extends MessageHandler {
     @Override
