+++ src/main/java/ru/hh/resumebuilderbot/question/generator/QuestionGeneratorByNumber.java
@@ -2,7 +2,7 @@ package ru.hh.resumebuilderbot.question.generator;
 
 import ru.hh.resumebuilderbot.ChatId;
 import ru.hh.resumebuilderbot.Question;
-import ru.hh.resumebuilderbot.QuestionsStorage;
+import ru.hh.resumebuilderbot.questions.storage.QuestionsStorage;
 
 public class QuestionGeneratorByNumber implements QuestionGenerator {
